Conflicts:
	simplesin.cpp

+++ simplesin.cpp
@@ -1,5 +1,5 @@
-
 double f(uint64_t t, double in)
 {
-   return sin(t / SampleRate * 440 * 2 * M_PI);
+   return sin((double)t / SampleRate * 440 * 2 * M_PI);
 }
+
